Implementation of delayed lowering transformations (PART 1)
This is the initial implementation of Leo's change on
delayed lowering transformations.

The implementation is based on building a queue of callbacks and
arguments (technically closures) that define what transformations to
perform. The treeLowering optimization will then consist of a pass that
finds the trees to be lowered and records them in the queue followed by
a pass (of the queue) that actually does the lowering transformations.

Authored-by: Leonardo Banderali <leonardo2718@protonmail.com>

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>