Do not trust nullable arrayclass type in VP
Update isUnreliableSignatureType to check if the class
is an array class. If the class is an array class and
its component class is value type and the array class
is not null-restricted array, the type is not reliable
because it can either be a nullable array class type
or a null-restricted array type.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>