Clean up references to old opcodes
The old opcodes had the letter i as a prefix if the operation
was indirect. Those opcodes were renamed to use the letter
as a suffix instead.

This commit cleans up the references to the old opcodes,
iiload/iistore ilload/ilstore iaload/iastore ibload/ibstore,
to iloadi/istorei lloadi/lstorei aloadi/astorei bloadi/bstorei
in the code comments and trace messages

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>