JProfiling fix to avoid too many internal pointers
Jprofiling adds control flow to the IL as part of the lowering of
the calls to profile values. This control flow splits blocks, and
in the process creates temps for values commoned across the split
point. If there are enough internal pointers commoned in the method,
the lower may end up creating more than the max number of internal
pointers allowed, thus causing the compile to fail. This commit
keeps track of how many internal pointers are created and avoids
lowering any more calls once we have reached a certain threshold.
Also added an env var to control this threshold for experimenting
in the future.

Signed-off-by: Vijay Sundaresan <vijaysun@ca.ibm.com>