Check Control Flow Guard on Windows
If Control Flow Guard (CFG) is enabled in Windows, set
TR_NoResumableTrapHandler in the JIT.

If CFG is set, the JIT will not resume exception but
instead it goes to the unexpected exception path error
handling so that the diagnostic logs will be dumped.

This change relies on API GetProcessMitigationPolicy.
Its minimum supported client version is Windows 8 and
its minimum supported server version is Windows Server 2012.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>